src/backend/requirements.txt
# FastAPI and Core Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic[email]==2.5.0
pydantic-settings==2.1.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-oauth2==1.1.1
authlib==1.2.1

# Database
sqlalchemy==2.0.23
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Caching & Background Tasks
redis==5.0.1
celery==5.3.4
aioredis==2.0.1

# External Integrations
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0
websockets==12.0

# GraphQL
strawberry-graphql[fastapi]==0.215.1
graphene==3.3
ariadne==0.21.0

# File Processing
python-magic==0.4.27
pillow==10.1.0
openpyxl==3.1.2
pandas==2.1.4
PyPDF2==3.0.1

# Monitoring & Logging
prometheus-client==0.19.0
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0